Move ServiceBuilderExt to layers
